device-simulator/requirements.txt
# HTTP client for API calls
requests==2.31.0

# Data generation and manipulation
faker==20.1.0
numpy==1.24.3

# Async support
aiohttp==3.9.1

# Configuration
python-decouple==3.8

# Logging
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
